invoice DTO BE validation
